pyproject.toml
[project]
name = "patientscheduler"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0",
    "django-filter>=25.2",
    "djangorestframework>=3.16.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "pytest>=9.0.2",
    "pytest-django>=4.11.1",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "clinicapp.settings"
testpaths = ["tests"]
python_files = ["test_*.py"]
